# Add JsonWriteObject perf tests and cleanup use of var