--- a/tests/Benchmarks/System.Text.JsonLab/JsonParserEnumeratePerf.cs
+++ b/tests/Benchmarks/System.Text.JsonLab/JsonParserEnumeratePerf.cs
@@ -17,17 +17,21 @@ namespace System.Text.JsonLab.Benchmarks
         private MemoryStream _stream;
         private StreamReader _reader;
 
+        private const int ArrayLength = 300; // We know Json400KB has a 300 length array within the payload.
+
+        private const int IterationCount = 1000;
+
         [GlobalSetup]
         public void Setup()
         {
             string jsonString = JsonStrings.Json400KB;
 
             // Remove all formatting/indendation
-            using (JsonTextReader jsonReader = new JsonTextReader(new StringReader(jsonString)))
+            using (var jsonReader = new JsonTextReader(new StringReader(jsonString)))
             {
                 JToken obj = JToken.ReadFrom(jsonReader);
                 var stringWriter = new StringWriter();
-                using (JsonTextWriter jsonWriter = new JsonTextWriter(stringWriter))
+                using (var jsonWriter = new JsonTextWriter(stringWriter))
                 {
                     obj.WriteTo(jsonWriter);
                     jsonString = stringWriter.ToString();
@@ -40,14 +44,14 @@ namespace System.Text.JsonLab.Benchmarks
         }
 
         [Benchmark(Baseline = true)]
-        public void ParseNewtonsoftEnumerate()
+        public void ParseNewtonsoftEnumerateArray()
         {
             _stream.Seek(0, SeekOrigin.Begin);
-            using (JsonTextReader jsonReader = new JsonTextReader(_reader))
+            using (var jsonReader = new JsonTextReader(_reader))
             {
                 JToken obj = JToken.ReadFrom(jsonReader);
 
-                for (int j = 0; j < 100; j++)
+                for (int j = 0; j < IterationCount; j++)
                 {
                     foreach (JToken token in obj)
                     {
@@ -57,16 +61,16 @@ namespace System.Text.JsonLab.Benchmarks
         }
 
         [Benchmark]
-        public void ParseNewtonsoftEnumerateReverse()
+        public void ParseNewtonsoftEnumerateArrayReverse()
         {
             _stream.Seek(0, SeekOrigin.Begin);
-            using (JsonTextReader jsonReader = new JsonTextReader(_reader))
+            using (var jsonReader = new JsonTextReader(_reader))
             {
                 JToken obj = JToken.ReadFrom(jsonReader);
 
-                for (int j = 0; j < 100; j++)
+                for (int j = 0; j < IterationCount; j++)
                 {
-                    for (int i = 299; i >= 0; i--)
+                    for (int i = ArrayLength - 1; i >= 0; i--)
                     {
                         JToken token = obj[i];
                     }
@@ -75,16 +79,16 @@ namespace System.Text.JsonLab.Benchmarks
         }
 
         [Benchmark]
-        public void ParseNewtonsoftFirst()
+        public void ParseNewtonsoftArrayFirst()
         {
             _stream.Seek(0, SeekOrigin.Begin);
-            using (JsonTextReader jsonReader = new JsonTextReader(_reader))
+            using (var jsonReader = new JsonTextReader(_reader))
             {
                 JToken obj = JToken.ReadFrom(jsonReader);
 
-                for (int j = 0; j < 100; j++)
+                for (int j = 0; j < IterationCount; j++)
                 {
-                    for (int i = 0; i < 300; i++)
+                    for (int i = 0; i < ArrayLength; i++)
                     {
                         JToken token = obj[0];
                     }
@@ -93,29 +97,46 @@ namespace System.Text.JsonLab.Benchmarks
         }
 
         [Benchmark]
-        public void ParseNewtonsoftMiddle()
+        public void ParseNewtonsoftArrayMiddle()
         {
             _stream.Seek(0, SeekOrigin.Begin);
-            using (JsonTextReader jsonReader = new JsonTextReader(_reader))
+            using (var jsonReader = new JsonTextReader(_reader))
             {
                 JToken obj = JToken.ReadFrom(jsonReader);
 
-                for (int j = 0; j < 100; j++)
+                for (int j = 0; j < IterationCount; j++)
                 {
-                    for (int i = 0; i < 300; i++)
+                    for (int i = 0; i < ArrayLength; i++)
                     {
-                        JToken token = obj[150];
+                        JToken token = obj[ArrayLength / 2];
                     }
                 }
             }
         }
 
         [Benchmark]
-        public void ParseSystemTextJsonLabEnumerate()
+        public void ParseNewtonsoftEnumerate()
+        {
+            _stream.Seek(0, SeekOrigin.Begin);
+            using (var jsonReader = new JsonTextReader(_reader))
+            {
+                JToken obj = JToken.ReadFrom(jsonReader);
+                JToken withinArray = obj[0];
+                for (int j = 0; j < IterationCount; j++)
+                {
+                    foreach (JToken childObject in withinArray)
+                    {
+                    }
+                }
+            }
+        }
+
+        [Benchmark]
+        public void ParseSystemTextJsonLabEnumerateArray()
         {
             JsonObject obj = JsonObject.Parse(_dataUtf8);
 
-            for (int j = 0; j < 100; j++)
+            for (int j = 0; j < IterationCount; j++)
             {
                 for (int i = 0; i < obj.ArrayLength; i++)
                 {
@@ -127,11 +148,11 @@ namespace System.Text.JsonLab.Benchmarks
         }
 
         [Benchmark]
-        public void ParseSystemTextJsonLabEnumerateReverse()
+        public void ParseSystemTextJsonLabEnumerateArrayReverse()
         {
             JsonObject obj = JsonObject.Parse(_dataUtf8);
 
-            for (int j = 0; j < 100; j++)
+            for (int j = 0; j < IterationCount; j++)
             {
                 for (int i = obj.ArrayLength - 1; i >= 0; i--)
                 {
@@ -143,11 +164,11 @@ namespace System.Text.JsonLab.Benchmarks
         }
 
         [Benchmark]
-        public void ParseSystemTextJsonLabFirst()
+        public void ParseSystemTextJsonLabArrayFirst()
         {
             JsonObject obj = JsonObject.Parse(_dataUtf8);
 
-            for (int j = 0; j < 100; j++)
+            for (int j = 0; j < IterationCount; j++)
             {
                 for (int i = 0; i < obj.ArrayLength; i++)
                 {
@@ -159,15 +180,31 @@ namespace System.Text.JsonLab.Benchmarks
         }
 
         [Benchmark]
-        public void ParseSystemTextJsonLabMiddle()
+        public void ParseSystemTextJsonLabArrayMiddle()
         {
             JsonObject obj = JsonObject.Parse(_dataUtf8);
 
-            for (int j = 0; j < 100; j++)
+            for (int j = 0; j < IterationCount; j++)
             {
                 for (int i = 0; i < obj.ArrayLength; i++)
                 {
-                    JsonObject withinArray = obj[150];
+                    JsonObject withinArray = obj[obj.ArrayLength / 2];
+                }
+            }
+
+            obj.Dispose();
+        }
+
+        [Benchmark]
+        public void ParseSystemTextJsonLabEnumerate()
+        {
+            JsonObject obj = JsonObject.Parse(_dataUtf8);
+            JsonObject withinArray = obj[0];
+
+            for (int j = 0; j < IterationCount; j++)
+            {
+                foreach (JsonObject childObject in withinArray)
+                {
                 }
             }