@ahsonkhan 2018-09-06
Adding foreach enumerator for arrays to benchmark.

``` ini

BenchmarkDotNet=v0.10.14.683-nightly, OS=Windows 10.0.17738
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-20180720-2
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT


```
|                                      Method |      Mean |     Error |    StdDev | Scaled | ScaledSD |    Gen 0 |    Gen 1 | Allocated |
|-------------------------------------------- |----------:|----------:|----------:|-------:|---------:|---------:|---------:|----------:|
|               ParseNewtonsoftEnumerateArray |  13.15 ms | 0.4755 ms | 1.3720 ms |   1.00 |     0.00 | 750.0000 | 375.0000 | 4761824 B |
|        ParseSystemTextJsonLabEnumerateArray | 174.08 ms | 3.4224 ms | 3.6619 ms |  13.37 |     1.35 |        - |        - |       0 B |
| ParseSystemTextJsonLabEnumerateForeachArray |  25.16 ms | 0.4977 ms | 0.6471 ms |   1.93 |     0.20 |        - |        - |       0 B |